requirements.txt
discord.py>=2.0
flask>=2.0
python-dotenv
requests
beautifulsoup4
apscheduler
aiohttp